// sdks/node-ts/src/apache_beam/proto/beam_fn_api.ts
export interface ProcessBundleDescriptor {
    /**
     * (Required) A pipeline level unique id which can be used as a reference to
     * refer to this.
     *
     * @generated from protobuf field: string id = 1;
     */
    id: string;
    /**
     * (Required) A map from pipeline-scoped id to PTransform.
     *
     * @generated from protobuf field: map<string, org.apache.beam.model.pipeline.v1.PTransform> transforms = 2;
     */
    transforms: {
        [key: string]: PTransform;
    };
    /**
     * (Required) A map from pipeline-scoped id to PCollection.
     *
     * @generated from protobuf field: map<string, org.apache.beam.model.pipeline.v1.PCollection> pcollections = 3;
     */
    pcollections: {
        [key: string]: PCollection;
    };
    /**
     * (Required) A map from pipeline-scoped id to WindowingStrategy.
     *
     * @generated from protobuf field: map<string, org.apache.beam.model.pipeline.v1.WindowingStrategy> windowing_strategies = 4;
     */
    windowingStrategies: {
        [key: string]: WindowingStrategy;
    };
    /**
     * (Required) A map from pipeline-scoped id to Coder.
     *
     * @generated from protobuf field: map<string, org.apache.beam.model.pipeline.v1.Coder> coders = 5;
     */
    coders: {
        [key: string]: Coder;
    };
    /**
     * (Required) A map from pipeline-scoped id to Environment.
     *
     * @generated from protobuf field: map<string, org.apache.beam.model.pipeline.v1.Environment> environments = 6;
     */
    environments: {
        [key: string]: Environment;
    };
    /**
     * A descriptor describing the end point to use for State API
     * calls. Required if the Runner intends to send remote references over the
     * data plane or if any of the transforms rely on user state or side inputs.
     *
     * @generated from protobuf field: org.apache.beam.model.pipeline.v1.ApiServiceDescriptor state_api_service_descriptor = 7;
     */
    stateApiServiceDescriptor?: ApiServiceDescriptor;
    /**
     * A descriptor describing the end point to use for Data API for user timers.
     * Required if the ProcessBundleDescriptor contains any transforms that have user timers.
     *
     * @generated from protobuf field: org.apache.beam.model.pipeline.v1.ApiServiceDescriptor timer_api_service_descriptor = 8;
     */
    timerApiServiceDescriptor?: ApiServiceDescriptor;
}
// ...
export const ProcessBundleDescriptor = new ProcessBundleDescriptor$Type();